Make adapt. hist. eq. method available to all methods
